backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "TEST BROO"
authors = ["4got10king <Aldavletshina77654@yandex.ru>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.8.2"
pydantic-settings = "^2.3.4"
alembic = "^1.13.2"
fastapi = "^0.111.1"
asyncpg = "^0.29.0"
aiohttp = "^3.10.1"
uvicorn = "^0.30.6"
greenlet = "^3.1.1"
alembic-postgresql-enum = "^1.3.0"
sqlalchemy = { extras = ["asyncio"], version = "^2.0.21" }
asyncio = "^3.4.3"
pytest-asyncio = "^0.24.0"
aiosqlite = "^0.20.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.4"
black = "^24.4.2"
pandas = "^2.2.2"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
exclude = [
    ".direnv",
    ".idea",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    "alembic"
]

# Same as Black.
line-length = 99
indent-width = 4

# Assume Python 3.10
target-version = "py310"

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.black]
line-length = 99
target-version = ["py310"]
skip-string-normalization = false

[tool.pytest.ini_options]
pythonpath = [
    ".", "src",
]
asyncio_mode="auto"